fixed seed credit card num

+++ server/db/models/user.js
@@ -39,7 +39,7 @@ const User = db.define('user', {
   },
   shippingAddress: Sequelize.STRING,
   billingAddress: Sequelize.STRING,
-  creditCardNumber: Sequelize.INTEGER,
+  creditCardNumber: Sequelize.BIGINT,
   creditCardCCV: Sequelize.INTEGER,
   creditCardExpDate: Sequelize.DATE
 })
